+++ ChangeLog.md
@@ -1,3 +1,14 @@
+# v1.2.2
+
+IMPROVEMENTS
+- Add REST API `POST /backup/tables/all`, fix `POST /backup/tables` to respect `CLICKHOUSE_SKIP_TABLES`
+
+BUG FIXES
+- fix [#297](https://github.com/AlexAkulov/clickhouse-backup/issues/297), properly restore tables where have fields with the same name as table name
+- fix [#298](https://github.com/AlexAkulov/clickhouse-backup/issues/298), properly create `system.backup_actions` and `system.backup_list` integration tables for ClickHouse before 21.1
+- fix [#303](https://github.com/AlexAkulov/clickhouse-backup/issues/303), ignore leading and trailing spaces in `skip_tables` and `--tables` parameters
+- fix [#292](https://github.com/AlexAkulov/clickhouse-backup/issues/292), lock clickhouse connection pool to single connection
+
 # v1.2.1
 
 IMPROVEMENTS